fix(wallet): move 404 page component out of pages dir to unbreak build
Next.js treats every module under src/pages as a route and the build's
page-type validator (.next/types/validator.ts) requires each to have a
default export. The colocated 404.spec.tsx had none, so `next build`
failed with "Property 'default' is missing ... PagesPageConfig".

Extract the component to src/components/shared/NotFoundPage following the
repo's thin-page convention (e.g. AuthPage, ApiKeysPage) and colocate its
spec there, leaving src/pages/404.tsx as a thin definePublicPage wrapper.
